some tests

+++ src/rosecommon/include/findclass.h
@@ -11,10 +11,10 @@ template <uint16_t T>
 struct find_class {};
 
 template <>
-struct find_class<ePacketType::PAKCS_LOGIN_REQ> { CliLoginReq type; };
+struct find_class<(uint16_t)ePacketType::PAKCS_LOGIN_REQ> { CliLoginReq type; };
 
 template <>
-struct find_class<ePacketType::PAKLC_LOGIN_REPLY> { SrvLoginReply type; };
+struct find_class<(uint16_t)ePacketType::PAKLC_LOGIN_REPLY> { SrvLoginReply type; };
 }
 
 #endif /* !_FINDCLASS_H_ */
